requirements.txt
djangorestframework
django-cors-headers
mysqlclient
Flask
djangorestframework-simplejwt